# shared/api-contract.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "静记 API Contract",
    "version": "1.0.0",
    "description": "三端共享的 API 契约文件，用于类型同步和文档生成"
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "username"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 },
                  "username": { "type": "string", "minLength": 2 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" },
                    "token": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "captchaId", "captchaCode"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" },
                  "captchaId": { "type": "string" },
                  "captchaCode": { "type": "string" },
                  "token": { "type": "string", "description": "可选：切换账号时传递旧 token" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" },
                    "token": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/profile": {
      "get": {
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "用户资料",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      },
      "put": {
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "avatar_url": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "登出成功" } }
      }
    },
    "/api/auth/change-password": {
      "post": {
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["oldPassword", "newPassword", "confirmPassword"],
                "properties": {
                  "oldPassword": { "type": "string" },
                  "newPassword": { "type": "string", "minLength": 6 },
                  "confirmPassword": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "密码修改成功" } }
      }
    },
    "/api/auth/current-book": {
      "put": {
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["book_id"],
                "properties": {
                  "book_id": { "type": "string", "format": "uuid" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "账本切换成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "book_id": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/transactions": {
      "get": {
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["income", "expense"] } },
          { "name": "category", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "startDate", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "endDate", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 10, "maximum": 100 } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string", "enum": ["amount", "date"] } },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "view", "in": "query", "schema": { "type": "string", "enum": ["own", "all"] } }
        ],
        "responses": {
          "200": {
            "description": "交易列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Transaction" } },
                    "total": { "type": "integer" },
                    "page": { "type": "integer" },
                    "pageSize": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount", "category", "type", "date"],
                "properties": {
                  "amount": { "type": "number", "minimum": 0.01 },
                  "category": { "type": "string", "format": "uuid" },
                  "type": { "type": "string", "enum": ["income", "expense"] },
                  "date": { "type": "string", "format": "date" },
                  "description": { "type": "string" },
                  "brand": { "type": "string" },
                  "location_name": { "type": "string" },
                  "latitude": { "type": "number" },
                  "longitude": { "type": "number" },
                  "poi_id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Transaction" }
              }
            }
          }
        }
      }
    },
...
    "/api/books": {
      "get": {
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "账本列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Book" }
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "maxLength": 100 },
                  "description": { "type": "string", "maxLength": 200 },
                  "icon": { "type": "string", "maxLength": 500 },
                  "icon_id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Book" }
              }
            }
          }
        }
      }
    },
...
    "/api/categories": {
      "get": {
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["income", "expense"] } }
        ],
        "responses": {
          "200": {
            "description": "分类列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Category" }
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "type"],
                "properties": {
                  "name": { "type": "string", "maxLength": 50 },
                  "icon": { "type": "string", "maxLength": 500 },
                  "type": { "type": "string", "enum": ["income", "expense"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          }
        }
      }
    },
    "/api/categories/{id}": {
      "put": {
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "maxLength": 50 },
                  "icon": { "type": "string", "maxLength": 500 },
                  "sort_order": { "type": "integer", "minimum": 0 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          }
        }
      },
      "delete": {
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "删除成功" } }
      }
    },
    "/api/categories/reorder": {
      "patch": {
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "required": ["id", "sort_order"],
                  "properties": {
                    "id": { "type": "string", "format": "uuid" },
                    "sort_order": { "type": "integer", "minimum": 0 }
                  }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "排序更新成功" } }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "username": { "type": "string" },
          "avatar_url": { "type": "string", "format": "uri" },
          "role": { "type": "string", "enum": ["user", "admin"] },
          "status": { "type": "string", "enum": ["active", "suspended", "deleted"] },
          "current_book_id": { "type": "string", "format": "uuid" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "amount": { "type": "number" },
          "category": { "type": "string", "format": "uuid" },
          "type": { "type": "string", "enum": ["income", "expense"] },
          "date": { "type": "string", "format": "date" },
          "description": { "type": "string" },
          "brand": { "type": "string" },
          "image_urls": { "type": "string" },
          "user_id": { "type": "string", "format": "uuid" },
          "book_id": { "type": "string", "format": "uuid" },
          "location_name": { "type": "string" },
          "latitude": { "type": "number" },
          "longitude": { "type": "number" },
          "poi_id": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Book": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "icon": { "type": "string" },
          "owner_id": { "type": "string", "format": "uuid" },
          "role": { "type": "string", "enum": ["owner", "member"] },
          "txn_count": { "type": "integer" },
          "member_count": { "type": "integer" },
          "is_archived": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "icon": { "type": "string" },
          "type": { "type": "string", "enum": ["income", "expense"] },
          "is_default": { "type": "boolean" },
          "sort_order": { "type": "integer" },
          "user_id": { "type": "string", "format": "uuid" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT-like-token"
      }
    }
  }
}